pyproject.toml
[project]
name = "yt_dlp_bot"
version = "0.1.0"
description = "Discord bot to auto download youtube streams"
authors = [
 {name = "xchgeaxeax"}
]
requires-python = ">=3.11"

dependencies = [
 "discord.py >= 2.5.2",
 "requests ~= 2.32.3",
 "yt-dlp (>=2025.8.11)", 
 "pydantic ~= 2.12.5",
 "httpx ~= 0.28.1",
 "streamlink >= 7.5.0"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-mock = "^3.12.0"
pytest-asyncio = "^0.23.5"
pytest-cov = "^5.0.0"
ruff = "^0.5.5"
